Fix benchmark compilation error and reduce batch size
- Fix type ambiguity in benchmark: i.to_le_bytes() -> (i as u64).to_le_bytes()
- Reduce GPU batch size from 4096 to 256 (8GB -> 512MB allocation)
- Large allocations were causing severe performance degradation

The 8GB buffer allocation was taking longer than actual computation!

+++ blakeout-gpu/benches/benchmark.rs
@@ -59,7 +59,7 @@ fn bench_comparison(c: &mut Criterion) {
             for i in 0..batch_size {
                 let mut hasher = Blakeout::new();
                 hasher.update(black_box(data));
-                hasher.update(&i.to_le_bytes());
+                hasher.update(&(i as u64).to_le_bytes());
                 black_box(hasher.result());
             }
         });
